backend/auth_service/requirements_simple.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
asyncpg==0.29.0
psycopg2-binary==2.9.10
pydantic[email]==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-jose[cryptography]==3.4.0
argon2-cffi==23.1.0
fastapi-csrf-protect==1.0.3
slowapi==0.1.9
python-multipart==0.0.6
